requirements.txt
flask~=2.3.2gunicorn~=20.1.0pytest~=7.4.0openai~=0.27.8beautifulsoup4~=4.12.2tiktoken~=0.4.0nltk~=3.8.1cassandra-driver~=3.28.0flask-wtf~=1.1.1https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp310-cp310-linux_x86_64.whlsentence-transformers~=2.2.2